requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
lxml==4.9.1
multitasking==0.0.11
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
schedule==1.1.0
six==1.16.0
tornado==6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
yfinance==0.1.77
